feat: restructure project to standard Python package layout
- Move source files to src/ directory
- Move tests to tests/ directory
- Update pyproject.toml for new structure
- Add Python 3.13 to CI matrix
- Update GitHub Actions for uv dependency groups
- Update README with new file paths
